+++ paifeaturestore-20230621/src/Models/GetInstanceResponseBody.php
@@ -4,12 +4,20 @@
 
 namespace AlibabaCloud\SDK\PaiFeatureStore\V20230621\Models;
 
+use AlibabaCloud\SDK\PaiFeatureStore\V20230621\Models\GetInstanceResponseBody\featureDBInfo;
 use AlibabaCloud\SDK\PaiFeatureStore\V20230621\Models\GetInstanceResponseBody\featureDBInstanceInfo;
 use AlibabaCloud\Tea\Model;
 
 class GetInstanceResponseBody extends Model
 {
     /**
+     * @var featureDBInfo
+     */
+    public $featureDBInfo;
+
+    /**
+     * @deprecated
+     *
      * @var featureDBInstanceInfo
      */
     public $featureDBInstanceInfo;
@@ -70,6 +78,7 @@ class GetInstanceResponseBody extends Model
      */
     public $type;
     protected $_name = [
+        'featureDBInfo' => 'FeatureDBInfo',
         'featureDBInstanceInfo' => 'FeatureDBInstanceInfo',
         'gmtCreateTime' => 'GmtCreateTime',
         'gmtModifiedTime' => 'GmtModifiedTime',
@@ -86,6 +95,9 @@ class GetInstanceResponseBody extends Model
     public function toMap()
     {
         $res = [];
+        if (null !== $this->featureDBInfo) {
+            $res['FeatureDBInfo'] = null !== $this->featureDBInfo ? $this->featureDBInfo->toMap() : null;
+        }
         if (null !== $this->featureDBInstanceInfo) {
             $res['FeatureDBInstanceInfo'] = null !== $this->featureDBInstanceInfo ? $this->featureDBInstanceInfo->toMap() : null;
         }
@@ -125,6 +137,9 @@ class GetInstanceResponseBody extends Model
     public static function fromMap($map = [])
     {
         $model = new self();
+        if (isset($map['FeatureDBInfo'])) {
+            $model->featureDBInfo = featureDBInfo::fromMap($map['FeatureDBInfo']);
+        }
         if (isset($map['FeatureDBInstanceInfo'])) {
             $model->featureDBInstanceInfo = featureDBInstanceInfo::fromMap($map['FeatureDBInstanceInfo']);
         }