+++ paifeaturestore-20230621/ChangeLog.md
@@ -1,3 +1,8 @@
+2025-07-17 Version: 4.0.2
+- Update API GetInstance: add response parameters Body.FeatureDBInfo.
+- Update API ListInstances: add response parameters Body.Instances.$.FeatureDBInfo.
+
+
 2025-07-16 Version: 4.0.1
 - Generated php 2023-06-21 for PaiFeatureStore.